Yet another restructure 🙃
