+++ core/src/com/csharp/game/screens/ui/screens/CreateLobbyScreen.java
@@ -1,0 +1,14 @@
+package com.csharp.game.screens.ui.screens;
+
+
+import com.csharp.game.RythemDiscord;
+
+public class CreateLobbyScreen extends UIScreen {
+
+    //textures and renderers
+
+    public CreateLobbyScreen(RythemDiscord game) {
+        super(game);
+    }
+
+}
