+++ core/src/com/csharp/game/screens/ui/screens/UIScreen.java
@@ -1,0 +1,92 @@
+package com.csharp.game.screens.ui.screens;
+
+import com.badlogic.gdx.Gdx;
+import com.badlogic.gdx.Screen;
+import com.badlogic.gdx.graphics.GL20;
+import com.badlogic.gdx.graphics.OrthographicCamera;
+import com.badlogic.gdx.scenes.scene2d.Stage;
+import com.badlogic.gdx.scenes.scene2d.ui.Skin;
+import com.badlogic.gdx.scenes.scene2d.ui.Table;
+import com.badlogic.gdx.utils.viewport.FillViewport;
+import com.badlogic.gdx.utils.viewport.Viewport;
+import com.csharp.game.RythemDiscord;
+
+public abstract class UIScreen implements Screen {
+    final RythemDiscord game;
+
+    //UI Items
+    public Skin skin;
+    public Table table;
+    public Stage stage;
+
+    //Camera & Viewport
+    private OrthographicCamera camera;
+    private Viewport viewport;
+
+    public UIScreen(RythemDiscord game) {
+        this.game = game;
+
+
+        this.skin = new Skin();
+        this.table = new Table();
+        this.stage = new Stage();
+        Gdx.input.setInputProcessor(stage);
+
+
+        this.camera = new OrthographicCamera();
+        this.viewport = new FillViewport(1000, 1000, this.camera);
+        this.camera.position.set(camera.viewportWidth / 2, camera.viewportHeight / 2, 0);
+    }
+
+
+    @Override
+    public void show() {
+
+    }
+
+    @Override
+    public void render(float delta) {
+        //Updating the camera
+        this.camera.update();
+        this.game.spriteBatch.setProjectionMatrix(this.camera.combined);
+
+        //Clear screen
+        Gdx.gl.glClear(GL20.GL_COLOR_BUFFER_BIT);
+
+        //render background
+        Gdx.gl.glClearColor(67 / 255f, 71 / 255f, 81 / 255f, 1);
+
+        //rendering the stage
+        stage.act(Math.min(Gdx.graphics.getDeltaTime(), 1 / 30f));
+        stage.draw();
+    }
+
+    @Override
+    public void resize(int width, int height) {
+        //Viewport and Camera update for SpriteBatch
+        this.viewport.update(width, height);
+        this.camera.position.set(camera.viewportWidth / 2, camera.viewportHeight / 2, 0);
+        //Viewport for UI Stage
+        this.stage.getViewport().update(width, height, true);
+    }
+
+    @Override
+    public void pause() {
+
+    }
+
+    @Override
+    public void resume() {
+
+    }
+
+    @Override
+    public void hide() {
+
+    }
+
+    @Override
+    public void dispose() {
+
+    }
+}
